+++ multinomial.go
@@ -5,11 +5,11 @@ import (
 )
 
 type Multinomial struct {
-  n int
+  n int64
   cat Categorical
 }
 
-func NewMultinomial(n int, p []float64) (mult Multinomial) {
+func NewMultinomial(n int64, p []float64) (mult Multinomial) {
   mult.n = n
   mult.cat = NewCategorical(p)
   return
@@ -17,7 +17,7 @@ func NewMultinomial(n int, p []float64) (mult Multinomial) {
 
 func (mult *Multinomial) Sample(r *rand.Rand) (s []float64) {
   s = make([]float64, len(mult.cat.p))
-  for i := 0; i < mult.n; i++ {
+  for i := int64(0); i < mult.n; i++ {
     idx := mult.cat.Sample(r)
     s[idx]++
   }
